adjust container & font styles
