+++ public/index.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="https://unpkg.com/bootstrap@3.3.7/dist/css/bootstrap.min.css">
 
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Basic</title>
+    <title>Expunge Colorado Screener</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
