+++ src/App.scss
@@ -5,45 +5,61 @@ $lightgray: #eee;
 $color-pale-purple: rgb(162, 139, 184);
 $color-dark-purple: rgb(133, 77, 185);
 
-.app {
-    display: flex;
-    width: 100vw;
-    align-items: center;
-    flex-direction: column;
+body {
     font-family: 'Quicksand', sans-serif;
 }
 
+.app {
+    margin: 0 auto;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    padding: 0 1em 2em 1em;
+    font-size: 16px;
+    width: 100%;
+
+    @media screen and (min-width: 800px) {
+        width: 80%;
+        max-width: 1000px;
+    }
+}
+
 h1 {
-    font-size: 1.75em;
+    font-size: 30px;
     font-weight: 600;
     margin-top: 0;
 }
 
+h5 {
+    font-size: 18px;
+    line-height: 1.5;
+}
+
 .logo {
-    width: 150px;
+    width: 200px;
 }
 
 .main {
-    width: 80%;
+    width: 100%;
 }
 
 .progress-container {
     position: relative;
     width: 100%;
+    margin: 1em 0;
 }
 
 .progress-background {
-    border-radius: 10px;
-    height: 3px;
     width: 100%;
+    border-radius: 4px;
+    height: 4px;
     background-color: $lightgray;
-    margin: 1em 0;
 }
 
 .progress-value {
     position: absolute;
-    border-radius: 10px;
-    height: 3px;
+    border-radius: 4px;
+    height: 4px;
     background-color: $gold;
 }
 
@@ -65,6 +81,7 @@ div.header {
 }
 
 div.container {
+    width: 100%;
     background-color: lightblue;
     padding-bottom: 1em;
     margin-bottom: 2em;
